requirements.txt
alembic==1.13.0
certifi==2023.11.17
charset-normalizer==3.3.2
greenlet==3.0.2
idna==3.6
Mako==1.3.0
MarkupSafe==2.1.3
psycopg2==2.9.9
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
slack-bolt==1.18.1
slack-sdk==3.26.1
SQLAlchemy==2.0.23
typing-extensions==4.9.0
urllib3==2.1.0
